refactor(store): move store related action, query keys and options to store folder
